feat: final

+++ app/src/main/java/com/example/vdk/data/database/entity/SensorEntity.kt
@@ -21,7 +21,7 @@ class SensorEntity(
         return Sensor(
             sound = this.sound.toDouble(),
             temperature = this.temperature,
-            time = this.time.toDouble(),
+            time = this.time,
             weight = this.weight
         )
     }
